+++ src/index.js
@@ -1,4 +1,6 @@
-module.exports = {
-  StorageProvider: require('./storage-provider'),
-  Plugin: require('./plugin'),
-};
+import StorageProvider from './storage-provider.js';
+
+export default StorageProvider;
+
+export { default as StorageProvider } from './storage-provider.js';
+export { default as Plugin } from './plugin.js';